+++ Assets/CommonCoreModules/World/WorldTypes.cs
@@ -169,7 +169,7 @@ namespace CommonCore.World
         IgnoreHitLocation = 1 << 8,
         AlwaysExtremeDeath = 1 << 9,
         NeverExtremeDeath = 1 << 10,
-        DamageOnlyShields = 1 << 10
+        DamageOnlyShields = 1 << 11
     }
 
     /// <summary>